RIO-8722: Large File Support in MP4 PDL and PPB

+++ engines/player/test/src/test_pv_player_engine_testset13.cpp
@@ -4194,7 +4194,7 @@ void pvplayer_async_test_genericplaypauserepositionresumetest::HandleInformation
         curpos.iPosUnit = PVPPBPOSUNIT_DATAPOSITION;
         if (iPlayer->GetCurrentPositionSync(curpos) == PVMFSuccess)
         {
-            fprintf(iTestMsgOutputFile, "Playback status(data position) %d bytes\n", curpos.iPosValue.datapos_value);
+            fprintf(iTestMsgOutputFile, "Playback status(data position) %d bytes\n", (int32)curpos.iPosValue.datapos_value);
         }
 
         fprintf(iTestMsgOutputFile, "------------------------------\n");
@@ -6240,7 +6240,7 @@ void pvplayer_async_test_genericopensetplaybackrangestartplaystoptest::HandleInf
         curpos.iPosUnit = PVPPBPOSUNIT_DATAPOSITION;
         if (iPlayer->GetCurrentPositionSync(curpos) == PVMFSuccess)
         {
-            fprintf(iTestMsgOutputFile, "Playback status(data position) %d bytes\n", curpos.iPosValue.datapos_value);
+            fprintf(iTestMsgOutputFile, "Playback status(data position) %d bytes\n", (int32)curpos.iPosValue.datapos_value);
         }
 
         fprintf(iTestMsgOutputFile, "------------------------------\n");
@@ -8305,7 +8305,7 @@ void pvplayer_async_test_genericopenplayrepositiontoendtest::HandleInformational
         curpos.iPosUnit = PVPPBPOSUNIT_DATAPOSITION;
         if (iPlayer->GetCurrentPositionSync(curpos) == PVMFSuccess)
         {
-            fprintf(iTestMsgOutputFile, "Playback status(data position) %d bytes\n", curpos.iPosValue.datapos_value);
+            fprintf(iTestMsgOutputFile, "Playback status(data position) %d bytes\n", (int32)curpos.iPosValue.datapos_value);
         }
 
         fprintf(iTestMsgOutputFile, "------------------------------\n");
@@ -17082,7 +17082,7 @@ void pvplayer_async_test_genericplaypauserepositionresumenwdisconnectcancelallte
         curpos.iPosUnit = PVPPBPOSUNIT_DATAPOSITION;
         if (iPlayer->GetCurrentPositionSync(curpos) == PVMFSuccess)
         {
-            fprintf(iTestMsgOutputFile, "Playback status(data position) %d bytes\n", curpos.iPosValue.datapos_value);
+            fprintf(iTestMsgOutputFile, "Playback status(data position) %d bytes\n", (int32)curpos.iPosValue.datapos_value);
         }
 
         fprintf(iTestMsgOutputFile, "------------------------------\n");
@@ -20463,7 +20463,7 @@ void pvplayer_async_test_genericopenplaygetmetadatatest::HandleInformationalEven
         curpos.iPosUnit = PVPPBPOSUNIT_DATAPOSITION;
         if (iPlayer->GetCurrentPositionSync(curpos) == PVMFSuccess)
         {
-            fprintf(iTestMsgOutputFile, "Playback status(data position) %d bytes\n", curpos.iPosValue.datapos_value);
+            fprintf(iTestMsgOutputFile, "Playback status(data position) %d bytes\n", (int32)curpos.iPosValue.datapos_value);
         }
 
         fprintf(iTestMsgOutputFile, "------------------------------\n");
@@ -22390,7 +22390,7 @@ void pvplayer_async_test_genericopengetmetadatapictest::HandleInformationalEvent
         curpos.iPosUnit = PVPPBPOSUNIT_DATAPOSITION;
         if (iPlayer->GetCurrentPositionSync(curpos) == PVMFSuccess)
         {
-            fprintf(iTestMsgOutputFile, "Playback status(data position) %d bytes\n", curpos.iPosValue.datapos_value);
+            fprintf(iTestMsgOutputFile, "Playback status(data position) %d bytes\n", (int32)curpos.iPosValue.datapos_value);
         }
 
         fprintf(iTestMsgOutputFile, "------------------------------\n");
