+++ fileformats/mp3/parser/src/mp3fileio.cpp
@@ -1,5 +1,5 @@
 /* ------------------------------------------------------------------
- * Copyright (C) 1998-2009 PacketVideo
+ * Copyright (C) 1998-2010 PacketVideo
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -278,5 +278,5 @@ uint8 MP3FileIO::peekNextByte(PVFile *fp)
 // return 0 if unknown
 uint32 MP3FileIO::getContentLength(PVFile *fp)
 {
-    return fp->GetContentLength();
+    return (uint32)fp->GetContentLength();
 }
